+++ checkpoint/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- added Zarr Version 3 allowing custom write and read chunk configurations
 
 ## [0.4.4] - 2023-11-29